Fixed a bug where the timer would run the onTimerComplete event even during a timer interrupt.
Fixed: Added a if statement to check if the timer has stopped.
This prevents the onTimerComplete event from being invoked when the timer is being told to be interrupted.

Note sooner or later we will implement an AwaitableCompletionSource for the timer. This will allow using  AwaitableCompletionSource.TrySetCanceled which has a built in timer interruption event in it.