Updated dependencies in the package json. This includes adding "com.unity.cinemachine": "3.1.2".
We added "com.unity.cinemachine": "3.1.2" to the package json because currently the version after that one has a bug that breaks camera confiners in several cases. This caused too many headaches and was easier to downgrade a patch for a fix to the bug than make a custom work around.