Added a check for when grounded and frozen to not update the horizontal velocity.
Note we don't freeze the vertical velocity at first because the character could start to be frozen in mid air and we want to ground them first than stop their velocity.

The vertical velocity will always be stopped during the IsGrounded checks so we don't need any extra checks here to make the vertical velocity 0.