Bugfix: The player input wasn't being disabled during pause, UI menus, and dialogue. Allowing some janky movement during these.
Updated:
Updated the Control keybindings and moved some stuff the the UI map and made a new map for GamePlay.

There is a Talk input action now inside of the UI map and also created a Gameplay map and put the pause input there.
Now we can disable the complete player input map without breaking interacting with dialogues and menu stuff.