First implementation for starting a new game and loading it from the main menu.
Added:
Implemented a event for SceneManager.sceneLoaded. This allows for calling events as a new scene is being loaded.
This allows us to set data when starting a new game or loading a game.
It also allows us to load a scene and do different things based on certain conditions.

Example usage: When starting a new game and loading the starting area scene we can set the player start room to the first room, have a cut scene be called and so forth.