Game Initialization Loop - Part One Database Foundation
Note: Some of the changes in this commit was done before the final design, so couple changes normally would of been put in part two.

This is the start of the new Game Initialization Loop. This loop controls the initial start up of the game when the first scene is loaded.
It makes sure the databases are loaded before any other core system that might rely on them. This runs first on game load no matter what the starting scene in the build is allowing for editor and build testing to make sure the required data is initialized before spawning or using and core system.