Major Milestone Reached: Created the now default base class for SF Shapes for low level physics 2D.
Added:
Created the SFShapeComponent that will replace the SFShapeBody.cs moving forward.
It is a simplified, but also improved component that does everything SFShapeBody, but better with a lot more safety checks and improved logic that allows for quickly creating new custom low level physics components with ease by inheriting from the class.

Created a SFRectangleShape component that will replace the RectangleBody2D classes.
This inherits from SFShapeComponent for a much better workflow and improved updating the shape properties in editor with instant in scene editing.

Made a file for OwnedShapes that act like a key value pair for PhysicShapes and OwnerKeys.
This will use purely NativeArray/NativeLists and span only types for high performance.